Refactor thumbnail classes out of repository project
