+++ protagonist.sln
@@ -57,6 +57,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Utils", "Utils", "{EEB2997F
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestData", "Utils\TestData\TestData.csproj", "{4561CCE1-5175-4A6A-9707-C2C4776C9D19}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Thumbs.Tests", "Thumbs.Tests\Thumbs.Tests.csproj", "{D8C11162-A3CF-475F-B182-5A115F6DE112}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -159,6 +161,10 @@ Global
 		{4561CCE1-5175-4A6A-9707-C2C4776C9D19}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4561CCE1-5175-4A6A-9707-C2C4776C9D19}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4561CCE1-5175-4A6A-9707-C2C4776C9D19}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D8C11162-A3CF-475F-B182-5A115F6DE112}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D8C11162-A3CF-475F-B182-5A115F6DE112}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D8C11162-A3CF-475F-B182-5A115F6DE112}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D8C11162-A3CF-475F-B182-5A115F6DE112}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -181,5 +187,6 @@ Global
 		{68876B84-0636-4573-A53A-926AB5C16231} = {89603DC1-C3AD-49C8-A412-D556AE6483D6}
 		{1E5EB337-60AA-4D86-BFFC-30FB747F9B25} = {89603DC1-C3AD-49C8-A412-D556AE6483D6}
 		{4561CCE1-5175-4A6A-9707-C2C4776C9D19} = {EEB2997F-35CD-4666-8E8F-9BF9CB5F46D9}
+		{D8C11162-A3CF-475F-B182-5A115F6DE112} = {89603DC1-C3AD-49C8-A412-D556AE6483D6}
 	EndGlobalSection
 EndGlobal
